+++ src/promptflow/promptflow/_cli/_pf_azure/_connection.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from dotenv import dotenv_values
 
-from promptflow._cli._params import add_param_connection_name, add_param_env, logging_params
+from promptflow._cli._params import add_param_connection_name, add_param_env, base_params
 from promptflow._cli._utils import _set_workspace_argument_for_subparsers, activate_action, get_client_for_cli
 from promptflow._sdk._constants import LOGGER_NAME
 from promptflow._utils.logger_utils import LoggerFactory
@@ -38,7 +38,7 @@ def add_connection_create(subparsers):
         add_param_connection_name,
         add_param_type,
         add_param_env,
-    ] + logging_params
+    ] + base_params
 
     activate_action(
         name="create",
@@ -56,7 +56,7 @@ def add_connection_get(subparsers):
         _set_workspace_argument_for_subparsers,
         add_param_connection_name,
         add_param_env,
-    ] + logging_params
+    ] + base_params
 
     activate_action(
         name="get",