+++ src/promptflow/tests/executor/e2etests/test_traces.py
@@ -1,4 +1,3 @@
-import pkg_resources
 from types import GeneratorType
 
 import pytest
@@ -7,7 +6,7 @@ from promptflow._utils.dataclass_serializer import serialize
 from promptflow.contracts.run_info import Status
 from promptflow.executor import FlowExecutor
 
-from ..utils import get_flow_sample_inputs, get_yaml_file
+from ..utils import get_yaml_file
 
 
 @pytest.mark.usefixtures("dev_connections")
@@ -47,37 +46,27 @@ class TestExecutorTraces:
 
         return get_trace
 
-    @pytest.mark.skipif(
-        pkg_resources.get_distribution("openai").version.startswith("1."),
-        reason="test needs to be upgraded to adapt to openai>=1.0.0",
+    def get_chat_input(stream):
+        return {
+            "question": "What is the capital of the United States of America?", "chat_history": [], "stream": stream
+        }
+
+    def get_comletion_input(stream):
+        return {"prompt": "What is the capital of the United States of America?", "stream": stream}
+
+    @pytest.mark.parametrize(
+        "flow_folder, inputs",
+        [
+            ("openai_chat_api_flow", get_chat_input(False)),
+            ("openai_chat_api_flow", get_chat_input(True)),
+            ("openai_completion_api_flow", get_comletion_input(False)),
+            ("openai_completion_api_flow", get_comletion_input(True)),
+            ("llm_tool", {"topic": "Hello", "stream": False}),
+            ("llm_tool", {"topic": "Hello", "stream": True}),
+        ]
     )
-    @pytest.mark.parametrize("flow_folder", ["openai_chat_api_flow", "openai_completion_api_flow"])
-    def test_executor_openai_api_flow(self, flow_folder, dev_connections):
+    def test_executor_openai_api_flow(self, flow_folder, inputs, dev_connections):
         executor = FlowExecutor.create(get_yaml_file(flow_folder), dev_connections)
-        inputs = get_flow_sample_inputs(flow_folder)
-        flow_result = executor.exec_line(inputs)
-
-        assert isinstance(flow_result.output, dict)
-        assert flow_result.run_info.status == Status.Completed
-        assert flow_result.run_info.api_calls is not None
-
-        assert "total_tokens" in flow_result.run_info.system_metrics
-        assert flow_result.run_info.system_metrics["total_tokens"] > 0
-
-        get_traced = False
-        for api_call in flow_result.run_info.api_calls:
-            get_traced = get_traced or self.validate_openai_apicall(serialize(api_call))
-
-        assert get_traced is True
-
-    @pytest.mark.skipif(
-        pkg_resources.get_distribution("openai").version.startswith("0."),
-        reason="Run tests for openai>=1.0.0",
-    )
-    @pytest.mark.parametrize("flow_folder", ["openai_v1_chat_api_flow", "openai_v1_completion_api_flow"])
-    def test_executor_openai_v1_api_flow(self, flow_folder, dev_connections):
-        executor = FlowExecutor.create(get_yaml_file(flow_folder), dev_connections)
-        inputs = get_flow_sample_inputs(flow_folder)
         flow_result = executor.exec_line(inputs)
 
         assert isinstance(flow_result.output, dict)