+++ src/promptflow/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Release History
 
-## 1.1.0 (Upcoming)
+## 1.2.0 (upcoming)
+
+### Bugs Fixed
+
+- [SDK/CLI] Removing telemetry warning when running commands.
+
+## 1.1.1 (2023.12.1)
+
+### Bugs Fixed
+
+- [SDK/CLI] Fix compatibility issue with `semantic-kernel==0.4.0.dev0` and `azure-ai-ml==1.12.0`.
+- [SDK/CLI] Add back workspace information in CLI telemetry.
+- [SDK/CLI] Disable the feature to customize user agent in CLI to avoid changes on operation context.
+- Fix openai metrics calculator to adapt openai v1.
+
+## 1.1.0 (2023.11.30)
 
 ### Features Added
 - Add `pfazure flow show/list` to show or list flows from Azure AI.
@@ -17,12 +32,14 @@
   - For remote workspace flow: `pfazure run create --flow azureml:<flow-name>`
   - For remote registry flow: `pfazure run create --flow azureml://registries/<registry-name>/models/<flow-name>/versions/<flow-version>`
 - Support set logging level via environment variable `PF_LOGGING_LEVEL`, valid values includes `CRITICAL`, `ERROR`, `WARNING`, `INFO`, `DEBUG`, default to `INFO`.
+- Remove openai version restrictions
 
 ### Bugs Fixed
 
 - [SDK/CLI] Fix node test with dict node input will raise "Required input(s) missing".
 - [SDK/CLI] Will use run name as display name when display name not specified (used flow folder name before).
 - [SDK/CLI] Fix pf flow build created unexpected layer of dist folder
+- [SDK/CLI] Fix deploy prompt flow: connections value may be none
 
 ### Improvements
 - Force 'az login' if using azureml connection provider in cli command.